+++ netlogo-gui/src/main/lab/gui/RunOptionsDialog.scala
@@ -5,7 +5,7 @@ package org.nlogo.lab.gui
 import org.nlogo.api.{Editable, Property}
 import org.nlogo.awt.UserCancelException
 import org.nlogo.core.{ I18N }
-import org.nlogo.lab.RunOptions
+import org.nlogo.api.RunOptions
 import org.nlogo.window.EditDialogFactoryInterface
 
 import java.io.File
